app/requirements.txt
flask
mysql-connector
requests
elementpath
flask_mysqldb
flask-cors